Fix REGRESS so all tests run on Mac OS X. (#1)
arnoldrobbins commented on 29 Aug 2018

The problem with Mac OS X is that by default the ulimit setting for core dumps is set to zero. This fixes REGRESS to make it be unlmited. Along the way, if a.out is not there it will run make and exit early if make fails.

I have tried to use very portable constructs in the shell script.
https://github.com/onetrueawk/awk/pull/13