Disable utf-8 for non-multibyte locales, such as C or POSIX.
This makes it possible to get the old awk behavior (where chars are
bytes) by setting LC_CTYPE to C or POSIX.  The value of MB_CUR_MAX
is cached since in many cases is actually a function.

No attempt is made to support arbitrary locales.  If the locale can
support multi-byte characters, utf-8 will be used.  If not, 8-bit
characters are used instead.