Discard excess function arguments
It's impossible for a node in a user-defined function to reference
arguments beyond those which are defined, so don't store them in the
arg lists. Evaluate them for side-effects and discard.

This change also plugs a tempcell leak in args[i] for i >= ndef.

===============> cat leak.sh
#/bin/sh

# Monitor awk's memory consumption while repeatedly calling a
# user-defined function with a couple of extra arguments.

${1:-./a.out} '
function f(x) { x }
BEGIN {
	mem = "ps -p $PPID -o rsz="
	system(mem)
	while (++i <= 100000) {
		f(i, i, i)
		if (i % 10000 == 0)
			system(mem)
	}
}
' 2>/dev/null
===============> paste <(./leak.sh ./master.out) <(./leak.sh ./a.out)
 2432	 2304
 3456	 2304
 4608	 2304
 5632	 2304
 6656	 2304
 7680	 2304
 8832	 2304
 9856	 2304
11008	 2304
12160	 2304
13312	 2304