Immediately propagate the conversion of an arg
There are at least two reasons for why the propagation of the
conversion of an uninitialized argument into an array cannot wait
until the function has finished.

The first is that otherwise any arguments in the active frame that
share the origin of the converted argument will persist for the
life of the function as independent copies of the uninitialized
value instead of referring to the new array.

The second is that the exit statement's longjmp entirely skips
the function epilogue (where the conversion had been handled),
opening the door to incorrect values in END blocks.

Instead of waiting for the epilogue, toarray immediately initiates
a scan of the stack frames to update those arguments that have the
same origin as the converted argument.

The argument list is now strictly of a single cell subtype, CCOPY. It
had been allowed to contain a mix of subtypes (CCOPY and CVAR). This
restriction simplifies the logic of the epilogue significantly.

The oargs list is no more. CCOPY cells use cnext to record their
origin (oargs recorded the parent, but the origin may be a more
distant ancestor).

Passing by reference had been done by setting the argument to the
address of the array's cell. This is not allowed under the new
regime. Instead it uses a pointer to the array's symbol table.
Which means that the table must be immobilzed or the stack frames
must be scanned upon relocation. Enter clearsymtab to empty arrays
without moving their symtabs.

This commit also fixes several other defects, among them a symbol
table leak (in the epilogue, unaware of each other, undetected
siblings would assign their private symbol tables to the same
parent), the use-after-free discovered by @millert in #157
(apologies for the delay in getting this out), and the following
use-after-free from a dangling name pointer (manifests only when
debugging is enabled):

----->$ cat dangling-nval.sh

awk=${1:-nawk}

valgrind --leak-check=full $awk -d '
function f(x) {
	g()
	print x
}

function g() {
	delete a["hi"]
}

BEGIN {
	a["hi"] = "world"
	f(a["hi"])
}'
----->$

For the moment, function arguments are anonymous. This change forced
a small modification to a test in T.misc.