Preserve user-defined function argument names
Use the new struct Function in much the same way as struct Array
(to make an object's internals reachable via Cell->sval).

While recording argument names, plug the leak of arglist nodes. In
the following program, f's arglist is leaked at the time of g's
definition:

	function f(a,b,c) {} function g() {}

The leak happens when varlist assigns to arglist in awkgram.y.

Note that though the parsing and defining of a function are now
limited only by available memory, invocations are still limited to
NARGS (an int).