Fix fnematch buffer use-after-free
The new fnematch simplifies the code base by using i, j, k pointers
to track the parsing.  Unfortunately, when `adjbuf` is called, the
pointers are can be invalidated because `adjbuf` uses `realloc`
internally and can move the passed `buf` to a new memory address.

This commit switches i, j, k to be indices relative to `buf` that
is updated on the fly.

Fixes the use-after-free on long input:

printf "%8192s\n" | tr " " "=" | ./a.out 'BEGIN{RS="th[^h]+"}{}'