Fix fnematch use-after-free on patbeg
A different use-after-free bug happens with patbeg;
patbeg can be set during the matching loop and does not get
updated with adjbuf.

To solve this, we add a new index, `patoff` that keeps the index
into buf until the end of the matching loop; on the loop exit,
`patbeg` will be updated from `patoff`.